+++ mangadex_downloader/cli/utils.py
@@ -43,6 +43,12 @@ def setup_network(args):
     # Setup delay requests (if set)
     Net.set_delay(args.delay_requests)
 
+    if args.dns_over_https:
+        Net.set_doh(args.dns_over_https)
+    
+    if args.timeout:
+        Net.set_timeout(args.timeout)
+
 def _keyboard_interrupt_handler(*args):
     print("Cleaning up...")
     # Downloader are not cleaned up