+++ docs/changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## UNRELEASED
+
+### New features
+
+- Added DNS-over-HTTPS support
+- Added ability to set timeout for each HTTP(s) requests
+
 ## v1.5.0
 
 ### New features